Fixed lint errors for details page
