+++ package.json
@@ -27,7 +27,7 @@
     "@types/react": "^17.0.37",
     "@types/react-helmet": "^6.1.4",
     "@types/react-router-dom": "5.1.2",
-    "@types/react-redux": "6.0.2",
+    "@types/react-redux": "7.1.24",
     "@types/webpack-dev-server": "^4.7.2",
     "@types/classnames": "^2.2.7",
     "@typescript-eslint/eslint-plugin": "^5.15.0",
