+++ src/components/helpers/stringHelpers.tsx
@@ -1,4 +1,4 @@
-import { GitProvider } from '../../types';
+import { GitProvider } from '../utils/types';
 
 export const gitUrlRegex =
   /^((((ssh|git|https?:?):\/\/:?)(([^\s@]+@|[^@]:?)[-\w.]+(:\d\d+:?)?(\/[-\w.~/?[\]!$&'()*+,;=:@%]*:?)?:?))|([^\s@]+@[-\w.]+:[-\w.~/?[\]!$&'()*+,;=:@%]*?:?))$/;
