+++ src/components/utils/types.ts
@@ -1,0 +1,17 @@
+import { K8sResourceCommon } from '@openshift-console/dynamic-plugin-sdk';
+
+export enum GitProvider {
+  GITHUB = 'github',
+  BITBUCKET = 'bitbucket',
+  GITLAB = 'gitlab',
+  UNSURE = 'other',
+  INVALID = '',
+}
+
+export type K8sResourceKind = K8sResourceCommon & {
+  spec?: {
+    [key: string]: any;
+  };
+  status?: { [key: string]: any };
+  data?: { [key: string]: any };
+};
