+++ src/index.d.ts
@@ -1,4 +1,4 @@
 declare module '*.png' {
-    const value: any;
-    export = value;
+  const value: any;
+  export = value;
 }
