--- src/types/git.ts
@@ -1,7 +1,0 @@
-export enum GitProvider {
-  GITHUB = 'github',
-  BITBUCKET = 'bitbucket',
-  GITLAB = 'gitlab',
-  UNSURE = 'other',
-  INVALID = '',
-}
